Don’t limit sparse regions to 2**32.
* paxlib/tar.h (struct sp_array):
Allow a sparse region to contain more than SIZE_MAX bytes;
this is quite possible on platforms with 32-bit size_t
and 64-bit off_t.

+++ paxlib/tar.h
@@ -262,7 +262,7 @@ enum archive_format
 struct sp_array
   {
     off_t offset;
-    size_t numbytes;
+    off_t numbytes;
   };
 
 union block
